+++ src/main/java/model/entity/Train.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.persistence.criteria.CriteriaBuilder;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,7 +13,7 @@ public class Train {
     private String id;
     private String routeId;
 
-    private Long compartmentFree;
-    private Long deluxeFree;
-    private Long berthFree;
+    private Integer compartmentFree;
+    private Integer deluxeFree;
+    private Integer berthFree;
 }
